backend/requirements.txt
# --- Web Framework ---
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1
aiofiles>=23.2.1

# --- AI & Vector Store ---
langchain>=0.1.13
langchain-openai>=0.1.1
langchain-chroma>=0.1.0
langchain-community>=0.0.29
pypdf>=4.1.0

# --- Networking ---
requests>=2.31.0
